Rename active stress and ionic model files with upper camel case
