+++ Code/Source/solver/ActiveStressODE.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
 // University of California, and others. SPDX-License-Identifier: BSD-3-Clause
 
-#include "active_stress_ode.h"
+#include "ActiveStressODE.h"
 
 void ActiveStressODE::read_model_specific_parameters(
     const ActiveStressModelParameters &params) {
