+++ Code/Source/solver/ActiveStressUniformUnsteady.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
 // University of California, and others. SPDX-License-Identifier: BSD-3-Clause
 
-#include "active_stress_uniform_unsteady.h"
+#include "ActiveStressUniformUnsteady.h"
 
 #include <fstream>
 #include <vector>
