+++ Code/Source/solver/IonicModel.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
 // University of California, and others. SPDX-License-Identifier: BSD-3-Clause
 
-#include "ionic_model.h"
+#include "IonicModel.h"
 
 #include "ComMod.h"
 #include "Parameters.h"
